pyproject.toml
[tool.poetry]
name = "voice"
version = "0.1.0"
description = ""
authors = ["ASWANTH C M <aswanthmanoj51@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
deepgram-sdk = "^3.2.7"
streamlit = "^1.34.0"
sseclient = "^0.0.27"
langchain-together = "^0.1.1"
pyaudio = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
